% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSession.R
\name{addSession}
\alias{addSession}
\title{Add Class-Session templet}
\usage{
addSession(filename = "session.Rmd")
}
\arguments{
\item{filename}{.Rmd file name want to be created. (Default: new.Rmd)}
}
\value{
A class-session Rmd templet.
}
\description{
This function will create a class-session Rmd file, which can
be edited further.
}
\examples{
addSession("new.Rmd")

}
