% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic}
\alias{basic}
\title{Basic templet}
\usage{
basic(filename = "new.Rmd")
}
\arguments{
\item{filename}{.Rmd file name want to be created. (Default: new.Rmd)}
}
\value{
A basic Rmd templet.
}
\description{
This function will create a basic Rmd file, which can
be edited further.
}
\examples{
basic("new.Rmd")

}
